requirements.txt
cython==0.29.1
h5py==2.8.0
nltk==3.4
numpy==1.15.4
Pillow==5.3.0
pycocotools==2.0.0
pyyaml==3.13.0
six==1.11.0
torch==1.0.0
tqdm==4.28.1

